--- CSETWebNg/src/app/services/navigation/workflow-classic/workflow-classic.component.html
@@ -1,1 +1,0 @@
-<p>workflow-classic works!</p>
