--- CSETWebNg/src/app/services/navigation/workflow-classic/workflow-classic.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-workflow-classic',
-  templateUrl: './workflow-classic.component.html',
-  styleUrls: ['./workflow-classic.component.scss']
-})
-export class WorkflowClassicComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
